--include/--exclude not working
grep has --include and --exclude options which I'd rather not have
to deal with in this program.  However, any --include or --exclude
options would go into "$@" which sucks because in order to work,
they need to be listed **before** user_exclude_files and
exclude_files.  Unfortunately, due to that whole thing where any
fe options have to come before grep options, using these options
(or any other grep options) such as to pass them to fe instead of
grep requires you to be aware of this.